modules/udac_person/requirements.txt
flask
flask_accepts
sqlalchemy
geoalchemy2
shapely
marshmallow_sqlalchemy
flask_cors
flask_sqlalchemy
grpcio
grpcio-tools
kafka-python
psycopg2
grpcio-health-checking